// frontend/src/api/revizit/model/waterReportKind.ts
/**
 * ReVIZit API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum WaterReportKind {
    PERCENTAGE = 'PERCENTAGE',
    SWAP = 'SWAP',
    REFILL = 'REFILL'
}